refactoring: docs for model/documenteditor/config
