+++ src/main/java/com/onlyoffice/model/documenteditor/config/document/ReferenceData.java
@@ -19,7 +19,7 @@
 package com.onlyoffice.model.documenteditor.config.document;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.onlyoffice.model.config.document.Document;
+import com.onlyoffice.model.documenteditor.config.Document;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
@@ -34,7 +34,7 @@ import lombok.Setter;
 public class ReferenceData {
     /**
      * Defines the unique document identifier used by the service to get a link to the file.
-     * It must not be changed when the document is edited and saved (i.e. it is not equal to the {@link Document#key document.key} parameter).
+     * It must not be changed when the document is edited and saved (i.e. it is not equal to the {@link Document#getKey()}  document.key} parameter).
      */
     private String fileKey;
